# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Biz Cards API",
        "description": "A simple API to manage business cards",
        "contact": {
            "email": "contact@yourapi.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/business-cards": {
            "get": {
                "tags": [
                    "Business Cards"
                ],
                "summary": "Display a listing of the business cards",
                "description": "Display a listing of the business cards.",
                "operationId": "59b2a6234dc7b632c6d538897b844bab",
                "responses": {
                    "200": {
                        "description": "List of business cards",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BusinessCard"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Business Cards"
                ],
                "summary": "Store a newly created business card",
                "description": "Store a newly created business card.",
                "operationId": "2bdf467824df3eda7b4e3017f6424a7f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "company",
                                    "title",
                                    "phone",
                                    "email",
                                    "address",
                                    "website"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "website": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Business card successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "company": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "phone": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "address": {
                                            "type": "string"
                                        },
                                        "website": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/business-cards/{id}": {
            "get": {
                "tags": [
                    "Business Cards"
                ],
                "summary": "Display the specified business card",
                "description": "Display the specified business card.",
                "operationId": "ec82048c0adb6048697735677e591b0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the business card",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business card details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessCard"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Business card not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Business Cards"
                ],
                "summary": "Update the specified business card",
                "description": "Update the specified business card.",
                "operationId": "69860c146046bdeb05f9d5432fd075ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the business card",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "company",
                                    "title",
                                    "phone",
                                    "email",
                                    "address",
                                    "website"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "website": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Business card successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BusinessCard"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Business card not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Business Cards"
                ],
                "summary": "Remove the specified business card",
                "description": "Remove the specified business card.",
                "operationId": "8f92e66e3d157b1133f0046fe041a80d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the business card",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Business card successfully deleted"
                    },
                    "404": {
                        "description": "Business card not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BusinessCard": {
                "required": [
                    "id",
                    "name",
                    "company",
                    "title",
                    "phone",
                    "email",
                    "address",
                    "website"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "company": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "address": {
                        "type": "string"
                    },
                    "website": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}